requirements.txt
Django
djangorestframework
python-decouple
psycopg2-binary
